Move DMA implementation to adc.rs
